+++ src/js/registry.cpp
@@ -440,8 +440,9 @@ namespace ccf::js
     {
       auto is_tx_committed =
         [this](ccf::View view, ccf::SeqNo seqno, std::string& error_reason) {
+          auto* current_consensus = get_consensus();
           return ccf::historical::is_tx_committed_v2(
-            consensus, view, seqno, error_reason);
+            current_consensus, view, seqno, error_reason);
         };
 
       ccf::historical::read_write_adapter_v4(